OffPolicyRLAlgorithm -> RLALgorithm
