+++ README.md
@@ -192,16 +192,16 @@ This repository supports the training of policies via two state-of-the-art
 off-policy RL algorithms: [TD3](https://arxiv.org/pdf/1802.09477.pdf) and 
 [SAC](https://arxiv.org/pdf/1801.01290.pdf).
 
-To train a policy using this algorithm, create a `OffPolicyRLAlgorithm` object 
+To train a policy using this algorithm, create a `RLAlgorithm` object 
 and execute the `learn` method, providing the algorithm the proper policy 
 along the process:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.fcnet.td3 import FeedForwardPolicy  # for TD3 algorithm
 
 # create the algorithm object
-alg = OffPolicyRLAlgorithm(policy=FeedForwardPolicy, env="AntGather")
+alg = RLAlgorithm(policy=FeedForwardPolicy, env="AntGather")
 
 # train the policy for the allotted number of timesteps
 alg.learn(total_timesteps=1000000)
@@ -262,9 +262,9 @@ To assign multiple CPUs/environments for a given training algorithm, set the
 `num_envs` term as seen below:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     # set num_envs as seen in the above figure
     num_envs=3,
@@ -356,11 +356,11 @@ like to train a fully connected network using the TD3 algorithm with a hidden
 size of [64, 64], this could be done as such:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.fcnet.td3 import FeedForwardPolicy  # for TD3 algorithm
 
 # create the algorithm object
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     policy=FeedForwardPolicy, 
     env="AntGather",
     policy_kwargs={
@@ -377,7 +377,7 @@ All `policy_kwargs` terms that are not specified are assigned default
 parameters. These default terms are available via the following command:
 
 ```python
-from hbaselines.algorithms.off_policy import FEEDFORWARD_PARAMS
+from hbaselines.algorithms.rl_algorithm import FEEDFORWARD_PARAMS
 print(FEEDFORWARD_PARAMS)
 ```
 
@@ -386,11 +386,11 @@ following commands:
 
 ```python
 # for TD3
-from hbaselines.algorithms.off_policy import TD3_PARAMS
+from hbaselines.algorithms.rl_algorithm import TD3_PARAMS
 print(TD3_PARAMS)
 
 # for SAC
-from hbaselines.algorithms.off_policy import SAC_PARAMS
+from hbaselines.algorithms.rl_algorithm import SAC_PARAMS
 print(SAC_PARAMS)
 ```
 
@@ -431,7 +431,7 @@ All `policy_kwargs` terms that are not specified are assigned default
 parameters. These default terms are available via the following command:
 
 ```python
-from hbaselines.algorithms.off_policy import GOAL_CONDITIONED_PARAMS
+from hbaselines.algorithms.rl_algorithm import GOAL_CONDITIONED_PARAMS
 print(GOAL_CONDITIONED_PARAMS)
 ```
 
@@ -440,11 +440,11 @@ default policy parameters can be found via the following commands:
 
 ```python
 # for TD3
-from hbaselines.algorithms.off_policy import TD3_PARAMS
+from hbaselines.algorithms.rl_algorithm import TD3_PARAMS
 print(TD3_PARAMS)
 
 # for SAC
-from hbaselines.algorithms.off_policy import SAC_PARAMS
+from hbaselines.algorithms.rl_algorithm import SAC_PARAMS
 print(SAC_PARAMS)
 ```
 
@@ -455,10 +455,10 @@ training by passing the term under the `meta_period` policy parameter.
 This can be assigned through the algorithm as follows:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy  # for TD3 algorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     policy=GoalConditionedPolicy,
     policy_kwargs={
@@ -535,10 +535,10 @@ set the `intrinsic_reward_type` attribute to the type of intrinsic reward you
 would like to use:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy  # for TD3 algorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     policy=GoalConditionedPolicy,
     policy_kwargs={
@@ -572,10 +572,10 @@ In order to use relative goals when training a hierarchical policy, set
 the `relative_goals` parameter to True:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy  # for TD3 algorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     policy=GoalConditionedPolicy,
     policy_kwargs={
@@ -599,10 +599,10 @@ when training a hierarchical policy, set the `off_policy_corrections` parameter
 to True:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy  # for TD3 algorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     policy=GoalConditionedPolicy,
     policy_kwargs={
@@ -722,10 +722,10 @@ In order to use hindsight action and goal transitions when training a
 hierarchical policy, set the `hindsight` parameter to True:
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy  # for TD3 algorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     policy=GoalConditionedPolicy,
     policy_kwargs={
@@ -756,10 +756,10 @@ term (<img src="/tex/fd8be73b54f5436a5cd2e73ba9b6bfa9.svg?invert_in_darkmode&san
 `cg_weights` term (see the example below).
 
 ```python
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 from hbaselines.goal_conditioned.td3 import GoalConditionedPolicy  # for TD3 algorithm
 
-alg = OffPolicyRLAlgorithm(
+alg = RLAlgorithm(
     ...,
     policy=GoalConditionedPolicy,
     policy_kwargs={
@@ -810,10 +810,10 @@ multi-agent algorithms:
   policy-specific attributes:
 
   ```python
-  from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+  from hbaselines.algorithms.rl_algorithm import RLAlgorithm
   from hbaselines.multiagent.td3 import MultiFeedForwardPolicy  # for TD3
   
-  alg = OffPolicyRLAlgorithm(
+  alg = RLAlgorithm(
       policy=MultiFeedForwardPolicy,
       env="...",  # replace with an appropriate environment
       policy_kwargs={}
@@ -831,10 +831,10 @@ multi-agent algorithms:
   attribute to True:
   
   ```python
-  from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+  from hbaselines.algorithms.rl_algorithm import RLAlgorithm
   from hbaselines.multiagent.td3 import MultiFeedForwardPolicy  # for TD3
   
-  alg = OffPolicyRLAlgorithm(
+  alg = RLAlgorithm(
       policy=MultiFeedForwardPolicy,
       env="...",  # replace with an appropriate environment
       policy_kwargs={
@@ -850,10 +850,10 @@ multi-agent algorithms:
   learners, algorithm, set the `maddpg` attribute to True:
   
   ```python
-  from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+  from hbaselines.algorithms.rl_algorithm import RLAlgorithm
   from hbaselines.multiagent.td3 import MultiFeedForwardPolicy  # for TD3
   
-  alg = OffPolicyRLAlgorithm(
+  alg = RLAlgorithm(
       policy=MultiFeedForwardPolicy,
       env="...",  # replace with an appropriate environment
       policy_kwargs={
