+++ experiments/run_multi_hrl.py
@@ -6,7 +6,7 @@ import sys
 
 from hbaselines.utils.misc import ensure_dir
 from hbaselines.utils.train import parse_options, get_hyperparameters
-from hbaselines.algorithms import OffPolicyRLAlgorithm
+from hbaselines.algorithms import RLAlgorithm
 
 EXAMPLE_USAGE = \
     'python run_multi_hrl.py "multiagent-ring-v0" --total_steps 1e6'
@@ -55,7 +55,7 @@ def run_exp(env,
     """
     eval_env = env if evaluate else None
 
-    alg = OffPolicyRLAlgorithm(
+    alg = RLAlgorithm(
         policy=policy,
         env=env,
         eval_env=eval_env,
