+++ hbaselines/algorithms/__init__.py
@@ -1,4 +1,5 @@
 """Init script for the algorithms submodule."""
 from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+from hbaselines.algorithms.rl_algorithm import RLAlgorithm
 
-__all__ = ["OffPolicyRLAlgorithm"]
+__all__ = ["OffPolicyRLAlgorithm", "RLAlgorithm"]
