+++ hbaselines/multiagent/base.py
@@ -20,9 +20,9 @@ class MultiActorCriticPolicy(ActorCriticPolicy):
       To train a policy using independent learners, do not modify any
       policy-specific attributes:
 
-      >>> from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+      >>> from hbaselines.algorithms import RLAlgorithm
       >>>
-      >>> alg = OffPolicyRLAlgorithm(
+      >>> alg = RLAlgorithm(
       >>>     policy=MultiActorCriticPolicy,
       >>>     env="...",  # replace with an appropriate environment
       >>>     policy_kwargs={}
@@ -38,9 +38,9 @@ class MultiActorCriticPolicy(ActorCriticPolicy):
       To train a policy using the shared policy feature, set the `shared`
       attribute to True:
 
-      >>> from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+      >>> from hbaselines.algorithms import RLAlgorithm
       >>>
-      >>> alg = OffPolicyRLAlgorithm(
+      >>> alg = RLAlgorithm(
       >>>     policy=MultiActorCriticPolicy,
       >>>     env="...",  # replace with an appropriate environment
       >>>     policy_kwargs={
@@ -54,9 +54,9 @@ class MultiActorCriticPolicy(ActorCriticPolicy):
       To train a policy using their MADDPG variants as opposed to independent
       learners, algorithm, set the `maddpg` attribute to True:
 
-      >>> from hbaselines.algorithms.off_policy import OffPolicyRLAlgorithm
+      >>> from hbaselines.algorithms import RLAlgorithm
       >>>
-      >>> alg = OffPolicyRLAlgorithm(
+      >>> alg = RLAlgorithm(
       >>>     policy=MultiActorCriticPolicy,
       >>>     env="...",  # replace with an appropriate environment
       >>>     policy_kwargs={
