+++ tests/fast_tests/test_fcnet.py
@@ -9,8 +9,8 @@ from hbaselines.fcnet.td3 import FeedForwardPolicy as TD3FeedForwardPolicy
 from hbaselines.fcnet.sac import FeedForwardPolicy as SACFeedForwardPolicy
 from hbaselines.fcnet.imitation import FeedForwardPolicy \
     as ImitationFeedForwardPolicy
-from hbaselines.algorithms.off_policy import SAC_PARAMS, TD3_PARAMS
-from hbaselines.algorithms.off_policy import FEEDFORWARD_PARAMS
+from hbaselines.algorithms.rl_algorithm import SAC_PARAMS, TD3_PARAMS
+from hbaselines.algorithms.rl_algorithm import FEEDFORWARD_PARAMS
 from hbaselines.algorithms.dagger import FEEDFORWARD_PARAMS \
     as IMITATION_FEEDFORWARD_PARAMS
